fix: add support for runner in integration tests
