+++ terraform_runner/src/module.rs
@@ -1,13 +1,17 @@
 use env_common::DeploymentStatusHandler;
 use env_defs::{ApiInfraPayload, CloudProvider};
+use log::{error, info};
 use std::path::Path;
 
 use env_common::{get_module_download_url, interface::GenericCloudHandler};
 
-pub async fn download_module(s3_key: &String, destination: &str) -> Result<(), anyhow::Error> {
+pub async fn download_module(
+    handler: &GenericCloudHandler,
+    s3_key: &String,
+    destination: &str,
+) -> Result<(), anyhow::Error> {
     println!("Downloading module from {}...", s3_key);
 
-    let handler = GenericCloudHandler::default().await;
     let url = match get_module_download_url(&handler, s3_key).await {
         Ok(url) => url,
         Err(e) => {
@@ -36,17 +40,17 @@ pub async fn download_module(s3_key: &String, destination: &str) -> Result<(), a
 }
 
 pub async fn get_module(
+    handler: &GenericCloudHandler,
     payload: &ApiInfraPayload,
     status_handler: &mut DeploymentStatusHandler<'_>,
 ) -> Result<env_defs::ModuleResp, anyhow::Error> {
     let track = payload.module_track.clone();
-    let handler = GenericCloudHandler::default().await;
     match handler
         .get_module_version(&payload.module, &track, &payload.module_version)
         .await
     {
         Ok(module) => {
-            println!("Successfully fetched module: {:?}", module);
+            info!("Successfully fetched module: {:?}", module);
             if module.is_none() {
                 let error_text = "Module does not exist";
                 println!("{}", error_text);
@@ -63,7 +67,7 @@ pub async fn get_module(
             }
         }
         Err(e) => {
-            println!("Failed to get module: {:?}", e);
+            error!("Failed to get module: {:?}", e);
             let status = "failed_init".to_string();
             let error_text: String = e.to_string();
             status_handler.set_status(status);
